+++ sphinx_lua_ls/config.py
@@ -17,6 +17,10 @@ T = _t.TypeVar("T")
 A = _t.ParamSpec("A")
 
 
+KNOWN_BROKEN_LUA_LS_RELEASES = ["3.16.0"]
+KNOWN_BROKEN_EMMYLUA_RELEASES = []
+
+
 @dataclass
 class LuaDomainConfig:
     project_root: pathlib.Path